+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+- Fixed calling `PybricksHub.write()` methods.
+
 ## [1.2.0] - 2025-07-11
 
 ### Added